requirements.txt
# ==============================================================================
# Python dependencies for `kiteconect_nifty_atr`
#
# Notes:
# - This file is for **pip** packages only.
# - Cron/tmux are **system** packages (install via dnf/apt), not pip:
#   - cronie (crontab + crond)
#   - util-linux (flock)
#   - tmux (optional, for attach/detach when running manually)
#   - firefox (used by Selenium token generation on ARM64 in this project)
# - Recommended interpreter for this repo: the project venv at `.venv/bin/python`
# ==============================================================================

aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
attrs==25.3.0
autobahn==19.11.2
Automat==25.4.16
backtesting==0.6.4
blinker==1.9.0
bokeh==3.7.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dash==3.1.1
et_xmlfile==2.0.0
Flask==3.1.1
fonttools==4.58.4
frozenlist==1.7.0
h11==0.16.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
incremental==24.7.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiteconnect==5.0.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mplfinance==0.12.10b0
multidict==6.6.3
narwhals==1.45.0
nest-asyncio==1.6.0
numba==0.61.2
numpy
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas>=2.3.2
pandas_ta==0.4.71b0
pillow==11.3.0
plotly==6.2.0
pluggy==1.6.0
propcache==0.3.2
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyotp==2.9.0
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
retrying==1.4.0
selenium==4.34.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tabulate==0.9.0
tqdm==4.67.1
tornado==6.5.1
trio==0.30.0
trio-websocket==0.12.2
twilio==9.6.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
xyzservices==2025.4.0
yarl==1.20.1
zipp==3.23.0
zope.interface==7.2
waitress==3.0.0
